MINOR: http-ana: Perform L7 retries because of status codes in response analyser
L7 retries because of status codes are now performed in the response
analyser. This way, it is no longer required to handle L7 retries in
si_cs_recv(). It is also useless to set CF_READ_ERROR on the response
channel to be able to trigger such retries.

In addition, if no L7 retries are performed when the response is received,
the L7 buffer is immediately released. Before in this case, it was only
released with the stream.

(cherry picked from commit 1f08bffe0cfaaedfc3e022c7c3be1580c3ed68de)
[Cf: This one was backported to 2.4 to ease future backports]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>