BUG/MEDIUM: compression: Properly get the next block to iterate on payload
When a DATA block is compressed, or when the compression context is finished
on a TLR/EOT block, the next block used to loop on the HTX message must be
refreshed because a defragmentation may have occurred.

This bug was introduced when the EOM block was removed in 2.4. Thus, this
patch must be backported to 2.4.

(cherry picked from commit 402740c3ad9bc0d294d0c161e4904c06396a0791)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>