Revert "BUG/MINOR: opentracing: initialization after establishing daemon mode"
This reverts commit f2263435d71964d1aa3eb80df6464500696c0515.

This commit is unnecessary because although it solves the problem of using
the OpenTracing filter in daemon mode, it does not solve the same problem
if chroot is used.

The following commit related to the OpenTracing filter solves both problems
efficiently.

(cherry picked from commit 164247405451abb6b7644fc9927e184892257f10)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>