CLEANUP: l7-retries: do not test the buffer before calling b_alloc()
The return value is enough now to know if the allocation succeeded or
failed.

This cleanup was already pushed by Willy (f499f50) but a revert crushed
it. It may be backported to the 2.4 because the original patch was done on
this version.

(cherry picked from commit 0fe1864f7db556f99f3a715a8a383902ed36bfae)
Signed-off-by: Willy Tarreau <w@1wt.eu>